vrend: Enable enhanced layouts also for GLES hosts
Since we no longer emit enhanced layouts on the host and handle
overlapping IO values and arrays by mapping them to vec4 values
or arrays we can enable enhanced layouts in the guest.

Some of the piglits related to arb_enhanced_layouts still fail,
but these actually require ARB_vertex_attrib_64bit, which is
something we lie about when running on GLES hosts.

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>