shader: ignore usage mask when syncing layouts and searching IO slot
Since we no longer emit enhanced layouts IO slots can no longer
overlap and we don't have to check the usage mask when searching
for the IO slot.

This fixes picking up stale IO info that may be passed in with the
shader key when it is initialized using an unrelated prev shader stage
that might still be bound.

Fixes: 028542a202afe31fee94656f10f0005115a24ec8
   shader: factor out input search and simplify code

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>