vrend/shader: Partially handle TGSI_SEMANTIC_EDGEFLAG
Edge flags may be emitted for points and lines to indicate whether
a vertex belongs to a boundary edges or is a boundary point, and only
if the flag is enabled, the line or point should be drawn. We can emulate
this by emitting the edge flag as extra variable and discard the fragment
in the FS, if the variable is zero.

This fixes piglits that use th edge flag with points.

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>