tests: Don't use ck_assert when creating the context
Success of context creation is usually checked by the caller, so
fix the one case where this is not done, and then remove the tests.
With that we can use this context creation also outside the "check"
environment, e.g. for fuzzing.

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1248>