fuzzer: Add fuzzer with some initial state and resource creation
This fuzzer creates and runs an initial valid compute or gfx state and
then uses the fuzzer input to create commands that are run one-by-one.

For each fuzzing payload a new context is create, then the initial state is
created and the valid command stream is run, and then the fuzzing buffer is
used to create commands. The commands are patched to contain handles to valid
resources or objects that are created on the fly.

In order to limit the number of resources and objects the handles are translated
to five bit ids that are then used to lookup or create resource/object, from
resource and object tables.

When a payload is finished the resources and objects are destroyed and the context
is released.

v2: - use handle value as seed to create texture resources
    - update main loop to use uint32_t bases data
    - Fix a few WS errors
    - Fix a few command sizes
    - use different iterator ranges for SO targets and sampler views
    - Add a break at switch/default
    - SSBO and Atomic update handle correction
    - Copy destroy resource from testvirgl
    - Seed buffer creation with handle
    - Don't always make sure that a resoure or object handle exists because
      we also want to test error path when a resource or object doesn't exist.
    - remove superfluous assignment
    - Unify updating resource handles

v3: - use "header" when decoding command
    - Swap assignment and command copying to make the logic easier to understand
    - Wrap size in parethesis and static assert its value
    - Separate the command payload size from the command size to not drop a
      dword with each iteration
    - return 0 for END_TRANSFERS size
    - use MIN2 to set available command size

v4: - Add comment about why we prefer large batch sizes
    - drop unused array buffer_desc and related typedef

All reviews  Ryan Neph

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1248>