vrend: support linear gbm_bo blob resources for dGPU prime
Add support for creating gbm_bo backed blob resources that can be shared
between iGPU and a passthrough dGPU in the guest. With this, guest could
blit dGPU rendered data to host linear resource directly.

The resource is created in two phases:
1. PIPE_RESOURCE_CREATE allocates a large enough gbm_bo
2. PIPE_RESOURCE_SET_TYPE imports it into GL via mem_obj according to
   the gbm_bo fd

So the key change is to allow linear targets to keep gbm_bo and create a
GL texture backed by gbm_bo and let guest map the blob resource so that
guest/host can share this resource directly to support dGPU prime.

Signed-off-by: Julia Zhang <julia.zhang@amd.com>