vrend: free unclaimed resources when context is destroyed
vrend_renderer_pipe_resource_create creates a resource that is then
usually claimed with a call vrend_get_blob_pipe. However, if not all
resources created like this are claimed then the memory used to allocate
these has to be freed when the context is destroyed to avoid a memory
leak.

Closes: https://gitlab.freedesktop.org/virgl/virglrenderer/-/issues/513

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1310>