vrend: workaround non-conformant sampler view pattern
Works around a rendering artifact in clients that are presumed to be
relying on implementation-defined behavior when sampling the alpha
channel of imported textures without one. Do the simplest thing and pass
the implementation-defined behavior decision to the native host driver.

Signed-off-by: Ryan Neph <ryanneph@google.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1344>