vrend: Check whether we actually got a valid OpenGL/GLES context
When we run on hardware that doesn't support OpenGL 3.0 context
creation will fail, but so far we didn't test this and would
eventually crash when epoxy aborts because a function pointer
can't be resolved.

Bail out when no GL context was created and also issue an error
message in this case. Also issue an error message when
we fail because support for UBOs is missing.

v2: * Add flag VREND_USE_GLES and pass it to
      vrend_renderer_init
    * Emit error message in vrend_renderer_init when context
      creation fails (both Ryan)

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1355>