drm/amdgpu: Fix undefined behavior due to bugs in API structs
Some of the API structs had implicit padding or were sized differently
on 32-bit and 64-bit systems.  Add explicit padding to prevent that.
Additionally, bool cannot be used in the API, as only 0 and 1 are valid
values for it and so a guest can trigger undefined behavior by passing
something that is not 0 or 1.  Instead, use <stdint.h> types for which
all bit patterns are valid.

This breaks the 32-bit ABI by changing the layout of struct
amdgpu_ccmd_bo_query_info_rsp.  However, that ABI probably has no users.
It fixes 32-bit guests on 64-bit hosts by ensuring that struct layouts
are the same for both.

The protocol definition header now uses a GCC pragma to cause the build
to fail if any of the structs require padding.  This only applies to the
contents of the header, not to any header files that it includes or to
files that include it.

Signed-off-by: Demi Marie Obenour <demi@invisiblethingslab.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1383>