virglrenderer: introduce virgl_renderer_resource_map_fixed
Introduce virgl_renderer_resource_map_fixed, a variant of
virgl_renderer_resource_map with that allows requesting the mapping
of the BO into an specific address.

This function must be called with a valid address to be used in mmap().
The address is owned by the caller and must be released by its own
means.

Signed-off-by: Sergio Lopez <slp@sinrega.org>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1374>