drm/amdgpu: Always use a shadow buffer
In theory, the shared memory used by the native context code should be
mapped read-only.  In practice, it is usually writable by the guest, so
the renderers must be secure against this.

To prevent TOCTOU, data races, and other bugs that might lead to
security problems, always use a shadow buffer for the response.  It is
possible to write secure command handlers without one, but a shadow
buffer makes them much less error-prone.

Also zero the shadow buffer, so that failing to initialize it (perhaps
in an error path) does not leak heap memory to the guest.

Signed-off-by: Demi Marie Obenour <demi@invisiblethingslab.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1383>