drm/amdgpu: Require 8-byte request and response alignment
Some protocol structures require 8-byte alignment, so each request and
the response must also be 8-byte aligned.  Otherwise, the code could
cast an insufficiently aligned pointer to a pointer to a struct
requiring 8-byte alignment, which is undefined behavior in C.

This also adds padding bytes to ensure that the size of each struct is a
multiple of 8.  These padding bytes are checked to be zero.

Signed-off-by: Demi Marie Obenour <demi@invisiblethingslab.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1383>