drm/amdgpu: Use flags instead of integers used as booleans
This replaces some integers that could only take the values 0 or 1 with
flags fields.  It also repurposes the high 32 bits of
struct amdgpu_ccmd_bo_va_op_req::flags for protocol use, as only the low
32 bits were actually passed to the kernel.

This is a highly backwards-incompatible change and so has been moved to
its own commit.

Signed-off-by: Demi Marie Obenour <demi@invisiblethingslab.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1383>