drm/amdgpu: only allow video processing with VIRGL_RENDERER_USE_VIDEO
Without VIRGL_RENDERER_USE_VIDEO guests shall not be able to submit to
the video hardware engines.  Future DRM renderers will also need to
honor this flag if they expose any media engines.

This requires that the media engines only be enumerated if
VIRGL_RENDERER_USE_VIDEO is set.  Otherwise, the guest would see the
engines but then receive an error when it tried to submit to them.  This
error would likely result in device loss.

Signed-off-by: Demi Marie Obenour <demi@invisiblethingslab.com>
Fixes: #535