drm/amdgpu: drop VA handling from GEM_NEW
Mixing the 2 actions was useful at first when guest commands
were immediately flushed.
Now that commands are properly buffered in Mesa, this doesn't
bring much and complexifies the Mesa code because it delays the
allocation until the VA operation.

So this commit drops the fields related to VA from GEM_NEW.

Signed-off-by: Pierre-Eric Pelloux-Prayer <pierre-eric.pelloux-prayer@amd.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1415>