proxy: print message when failing to pre-initialize a proxy context
Print a message when failing o pre-initialize a proxy context, for
example when failing to set up shared memory. This could make it easier
to spot problems when porting to other platforms.

For example a message like this could have helped me to debug a problem
with memfd_create when building using the android toolchain.

Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1430>