ci: build jobs trigger manual
From the jobs in the build stage, some of them are triggered when the pipeline
is created, using resources on the pipeline that may not be completed later.
It could be nice if the rules are adjusted to not trigger those jobs
automatically and wait for a user interaction or handled by ci-r'n'm.

Signed-off-by: Sergi Blanch Torne <sergi.blanch.torne@collabora.com>