drm: Split ENABLE_DRM into DRM and LIBDRM
With the addition of the asahi DRM native context renderer it was in
https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1541
suggested to use ENABLE_DRM instead to test for each DRM native context
renderer individually. This theoretically clashes with defining
ENABLE_DRM when libdrm is found. Existing use of `#if ENABLE_DRM`
appears to be DRM native context renderer specific so the only use of
the new ENABLE_LIBDRM is conditional inclusion of `xf86drm.h`.

Signed-off-by: Janne Grunau <j@jannau.net>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1541>