ci: Review rules
After the modification on !1526, when there are changes set manual jobs in
Marge pipelines that make Marge fail (see !1542#note_3034392).

The rule about when there are changes on some files should be evaluated when
there isn't a previous rule based on the environment that already defined if
the job should be included and how.

Signed-off-by: Sergi Blanch-Torne <sergi.blanch.torne@collabora.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1546>