vkr: force WC for blob mem mapping on dGPU driver
Some dGPU driver (ANV on Intel ARC) might unconditionally report a
coherent-cached mem type for the system memory heap. Across PCI bus,
the best we can do used to be WC, and that's still true in most cases.
CXL is quite new. Even on supported hardware, whether that works with
SLAT is unknown at this point. So we force WC whenever we see dGPU for
better compatibility and robustness.