vkr: emulate EXT_external_memory_dma_buf when unavailable
On macOS we use EXT_external_memory_metal for exporting/importing memory
buffers. However, guest Mesa uses dmabufs and assumes the host uses it as
well. Therefore, we need to "hide" any dmabuf extension and bits from
the host as well as "emulate" the existance of dmabuf bits when returning
results from host to guest.