proxy: introduce in-process rendering
Metal does not have the concept of sharing GPU resources between processes
so we have to do rendering in the same process that handles the commands.
When 'thread' render server worker is selected, we will now support init
with !VIRGL_RENDERER_RENDER_SERVER by creating a new thread in the host
process to act as the server.